// web/src/api/model/contractStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * RentGraph API
 * OpenAPI spec version: 0.1.0
 */

export type ContractStatus =
  (typeof ContractStatus)[keyof typeof ContractStatus];

export const ContractStatus = {
  uploaded: "uploaded",
  analyzing: "analyzing",
  done: "done",
  failed: "failed",
} as const;